% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaview_wrap.R
\name{aliview}
\alias{aliview}
\title{AliView: DNA sequences viewer}
\usage{
aliview(x, aliview.exec = "aliview")
}
\arguments{
\item{x}{DNA Sequences in a \code{DNAbin} object.}

\item{aliview.exec}{a character string giving the path of the program.}
}
\description{
This function uses AliView (Larsson, 2014) to visualize DNA sequences.
}
\references{
Larsson, A. (2014). AliView: a fast and lightweight alignment viewer and editor for large data sets. Bioinformatics30(22): 3276-3278.
}
