% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanID8}
\alias{cleanID8}
\title{Clean Database IDs from sequence names}
\usage{
cleanID8(x)
}
\arguments{
\item{x}{a vector of character strings to process.}
}
\value{
A vector of character strings without database IDs.
}
\description{
Clean Database IDs from sequence names (suffix separated by underscore).
}
