% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{countBC}
\alias{countBC}
\title{Count similar barcodes}
\usage{
countBC(file)
}
\arguments{
\item{file}{a character string giving the path of the Mothur file which the data are to be read from.}
}
\description{
This function counts the number of similar barcodes from a specific file generated by Mothur.
}
