% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diatobc-package.R
\docType{data}
\name{diatoms.seq}
\alias{diatoms.seq}
\title{Diatoms genetic sequences}
\format{an alignment in \code{DNAbin} format.}
\usage{
data(diatoms.seq)
}
\description{
An alignment of genetic sequences for two markers (18S and rbcl) for
548 diatom species and a non-diatom species ('Bolidomonas pacifica').
}
\keyword{datasets}
