% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diatobc-package.R
\docType{data}
\name{diatoms.tree}
\alias{diatoms.tree}
\title{Diatoms Phylogeny}
\format{a phylogenetic tree in \code{phylo} format.}
\usage{
data(diatoms.tree)
}
\description{
A phylogenetic tree of 548 diatom species.
}
\details{
The phylogeny has been reconstructed with RaxML and two genetic markers (18S and rbcl).
The tree is not rooted but can be rooted with 'Bolidomonas pacifica', a non-diatom species included in the phylogeny.
}
\keyword{datasets}
