% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_graphclust.R
\name{extendGraphClust}
\alias{extendGraphClust}
\title{Extend clusters to new species using a phylogenetic tree}
\usage{
extendGraphClust(clusters, target, pb = TRUE)
}
\arguments{
\item{clusters}{an object of class \code{graphclust}}

\item{target}{a \code{phylo} object}

\item{pb}{a logical. Should a progress bar be printed? (default TRUE).}
}
\value{
A vector giving the cluster of each tip.
}
\description{
This function affiliate new species to clusters of a \code{graphclust} object.
This is done using a phylogeny including both the species included in the \code{graphclust}
object and the set of new species to affiliate.
}
\details{
The affiliation algorithm is basic and works with phylogenetic distance as number of nodes between two tips.
It may produce \code{NA} values in case of ambiguity between two clusters.
Affliliation of a large number of new tips can be slow.
}
