% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_barcodes.R
\name{insertBC}
\alias{insertBC}
\title{Insert barcodes in an alignment}
\usage{
insertBC(bc, ref, region, align.method = "mafft", align.exec = NULL,
  threads = NULL, trim = FALSE)
}
\arguments{
\item{bc}{barcodes sequences. A \code{DNAbin} object or a path of a sequence file to be read in fasta format.}

\item{ref}{the reference alignment. A \code{DNAbin} object or a path of a sequence file to be read in fasta format.}

\item{region}{a vector of two integers giving the boundaries of the barcodes region within the reference alignment.}

\item{align.method}{a character string giving the method to use for alignment.
Available methods are "\code{mafft}" and "\code{muscle}".}

\item{align.exec}{a character string giving the path of the program if necessary.}

\item{threads}{an integer giving the number of threads to use. Available only with mafft.}

\item{trim}{logical. If \code{TRUE}, only the barcode region is returned.}
}
\value{
A \code{DNAbin} alignment.
}
\description{
This function inserts new barcodes into an alignment.
For this purpose, the barcode region of the reference alignment is re-aligned with the new set of barcodes.
}

