% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diatobc-package.R
\docType{data}
\name{IPSP68}
\alias{IPSP68}
\title{Diatoms IPS-P68 index}
\format{a list of two elements:
\describe{
  \item{clusters}{a \code{graphclust} object (package \pkg{phylosignal}) giving the 68 clusters of the index.}
  \item{scores}{IPSS (sensitivity) and IPSV (variability) for each cluster}
}}
\usage{
data(IPSP68)
}
\description{
This dataset contains all the necessary data to compute the IPS-P index based on 68 clusters as described in Keck et al. (2016).
}
\references{
Keck F., Rimet F., Franc A. & Bouchez A. (2016) Linking Diatoms Ecological Preferences To Phylogeny: New Perspectives for Aquatic Ecosystems Bioassessment.
}
\keyword{datasets}
