% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jplace_expand.R
\name{jplaceExpand}
\alias{jplaceExpand}
\title{Expanded \code{.jplace} files.}
\usage{
jplaceExpand(input, output, round.brlen = 6, likw.max = NA)
}
\arguments{
\item{input}{Input file, absolute path or relative to the current working directory (a jplace file).}

\item{output}{Ouput file, absolute path or relative to the current working directory  (a jplace file).
If set to \code{NULL}, file content is returned in R directly as a character string.}

\item{round.brlen}{numeric value. Number of digits to round branch lengths of the tree.
If \code{NA}, the tree is not modified.}

\item{likw.max}{numeric value. If not \code{NA}, the replacements with an inferior likelihood weight are dropped.}
}
\description{
This is an utility function to expand \code{.jplace} files generated by \code{RAxML} and \code{pplacer}.
}
\details{
This utility function is used to generate \code{.jplace} files which can be used with the function \code{tog}
of \code{guppy} to get phylogenetic trees.
}
\references{
Matsen F.A., Hoffman N.G., Gallagher A. & Stamatakis A. (2012) A Format for Phylogenetic Placements. PLoS ONE 7, e31009.
}
