% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafft_wrap.R
\name{mafft2}
\alias{mafft2}
\title{Interface with MAFFT}
\usage{
mafft2(x, y, add, method = "auto", maxiterate = 0, op = 1.53, ep = 0,
  gt, options, path, quiet)
}
\description{
This is a slightly modification of the \code{\link[ips]{mafft}} function from the package \pkg{ips} to allow multi-threading.
}
