% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{one_seq_one_spec}
\alias{one_seq_one_spec}
\title{Keep on sequence per species}
\usage{
one_seq_one_spec(x, taxa = labels.DNAbin(x), excl.regex = "_sp\\\\.")
}
\arguments{
\item{x}{a \code{DNAbin} alignment}

\item{taxa}{the taxa list vector. Should be as long as the number of sequences.}

\item{excl.regex}{a regular expression to keep some taxa as is. Useful for undeterminated species.}
}
\value{
A \code{DNAbin} matrix.
}
\description{
Keep one sequence per species by consensus
}
