% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathd8_wrap.R
\name{pathd8}
\alias{pathd8}
\title{Tree relative datation with PATHd8}
\usage{
pathd8(phy, pathd8.exec)
}
\arguments{
\item{phy}{a phylogenetic tree (class \code{phylo}) to date.}

\item{pathd8.exec}{a character string giving the path of the program.}
}
\value{
An ultrametric tree of class \code{phylo}.
}
\description{
This function uses PATHd8 to date phylogenetic trees with the Mean Path Length algorithm.
}
\details{
This is a limited wrapper of the C program PATHd8 to perform tree datation in relative time
with the Mean Path Length algorithm. Wrapper for calibrated datation (D8) is not implemented.
}
\references{
\itemize{
 \item Britton, T., Anderson, C. L., Jacquet, D., Lundqvist, S., & Bremer, K. (2007). Estimating divergence times in large phylogenetic trees. Systematic biology, 56(5), 741-752.
 \item Britton, T., Oxelman, B., Vinnersten, A., & Bremer, K. (2002). Phylogenetic dating with confidence intervals using mean path lengths. Molecular phylogenetics and evolution, 24(1), 58-65.
}
}
