% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raxml_wrap.R
\name{raxmlBackbone}
\alias{raxmlBackbone}
\title{RAxML phylogeny with backbone tree.}
\usage{
raxmlBackbone(x, backbone, model, raxml.exec, nruns = 1, bs = NULL,
  part = NULL, threads = NULL, MLSCC = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{an alignment of sequences of class \code{DNAbin}.}

\item{backbone}{a phylogenetic tree (class \code{phylo}) to use as a constraint.
The tip labels must match with the appropriate DNA sequences labels. Use \code{NULL} for no constraint.}

\item{model}{a character string giving the substitution model to use (the -m option of RAxML).}

\item{raxml.exec}{a character string giving the path of the program.}

\item{nruns}{the number of independent ML searches (runs). This is ignored if bootstrapping is enabled.
Then the number of ML searches will be the number of bootstraps set with \code{bs} divided by 5.}

\item{bs}{a numeric giving the number of rapid bootstraps to perform.}

\item{part}{a vector of integers giving the boundaries of the different partitions, if relevant.
E.g. \code{c(1, 5355, 5356, 7633)} for a two genes partition.}

\item{threads}{the number of threads to use, if available.}

\item{MLSCC}{a logical indicating whether to use the ML search convergence criterion (the -D option of RAxML).
This is a shortcut in ML search which can improve significantly the speed for large phylogenies.}

\item{quiet}{a logical indicating whether RAxML warnings are printed on console (available only for RAxML v8+).}
}
\value{
A \code{list} of two elements:
\describe{
  \item{\code{tree}}{The best-scoring ML tree as a \code{phylo} object}
  \item{\code{info}}{The log file printed by RAxML.}
}
}
\description{
This function uses RAxML to reconstruct a tree given an alignment
and a backbone tree (the -g option of RAxML).
}
\details{
Since RAxML doesn't support special characters in taxa/sequence names,
they are deleted automatically with a warning.
Be warned that deletion of illegal characters may lead to ambiguities in taxa/sequence names.
}
\references{
Stamatakis A. (2014) RAxML Version 8: A tool for Phylogenetic Analysis and Post-Analysis of Large Phylogenies. Bioinformatics.
}
\seealso{
\code{\link[ips]{raxml}} in the package \pkg{ips} for a general RAxML interface,
\code{\link{raxmlEPA}} for RAxML Evolutionary Placement Algorithm.
}
