% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{raxmlPart}
\alias{raxmlPart}
\title{Partitions for RAxML}
\usage{
raxmlPart(part)
}
\arguments{
\item{part}{a vector of integers giving sequentially the partitions limits.}
}
\description{
This function takes a vector of nucleotide positions (delimiting partitions) and format it for RAxML.
}
