% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaview_wrap.R
\name{seaview}
\alias{seaview}
\title{SeaView: DNA sequences and phylogenetic tree viewer}
\usage{
seaview(x, seaview.exec = "seaview")
}
\arguments{
\item{x}{DNA Sequences in a \code{DNAbin} object or a phylogenetic tree (class \code{phylo}).}

\item{seaview.exec}{a character string giving the path of the program.}
}
\description{
This function uses SeaView (Gouy, Guindon & Gascuel, 2010) to visualize DNA sequences and phylogenetic trees.
}
\references{
Gouy M., Guindon S. & Gascuel O. (2010) SeaView version 4 :
a multiplatform graphical user interface for sequence alignment and phylogenetic tree building.
Molecular Biology and Evolution 27(2):221-224.
}
